chore: reorganized code and cleaned up comments
