+++ mock/src/App.tsx
@@ -1,11 +1,11 @@
-import REPL from './REPL'
+import REPL from './components/REPL'
 import './App.css';
 
 function App() {
   return (
     <div className="App">
       <p className="App-header">
-        React Example
+        <h1>Echo</h1>
       </p>
       <REPL />      
     </div>
