+++ mock/src/components/REPL.tsx
@@ -1,0 +1,18 @@
+import '../REPL.css';
+import { REPLInput } from './REPLInput';
+
+/* 
+  You'll want to expand this component (and others) for the sprints! Remember 
+  that you can pass "props" as function arguments. If you need to handle state 
+  at a higher level, just move up the hooks and pass the state/setter as a prop.
+  
+  This is a great top level component for the REPL. It's a good idea to have organize all components in a component folder.
+  You don't need to do that for this gearup.
+*/
+export default function REPL() {
+  return (
+    <div className="full-REPL">   
+      <REPLInput />
+    </div>
+  );
+}
