--- mock/styles/main.css
@@ -1,28 +1,0 @@
-/*
-  This is a reasonable starting point for the CSS in Sprint 1.
-  Positioning in CSS can sometimes be annoying to manage, so we've opted to 
-  give you this starter CSS. Feel free to modify it as you wish, though.
-*/
-
-.repl {
-    position:relative;
-    min-height:95vh;    
-}
-
-.repl-history {      
-    height:85vh;   
-    margin: auto;
-    display: block;
-    overflow: scroll;    
-}
-
-.repl-input {        
-    margin: auto;    
-    height:10vh;
-}
-.repl-command-box {
-    width: 90%;            
-    margin: auto;
-    display: block;
-}
-
